backend/requirements.txt
Flask>=2.0
flask-cors
Flask-SQLAlchemy>=2.5
psycopg2-binary
python-dotenv
alembic
Flask-Bcrypt
Flask-JWT-Extended
APScheduler>=3.8
python-dateutil>=2.8
pytz
gunicorn
eventlet
google-generativeai
google-cloud-storage
pydub
Pillow
gunicorn
apscheduler
assemblyai
pycryptodome
bcrypt
setuptools
flask-socketio